# Use Link Time Optimisation whenever possible