--- a/.github/workflows/osrm-backend.yml
+++ b/.github/workflows/osrm-backend.yml
@@ -191,57 +191,57 @@ jobs:
           #   OSRM_CONNECTION_RETRIES: 10
           #   OSRM_CONNECTION_EXP_BACKOFF_COEF: 1.5
 
-          - name: clang-15-release
-            continue-on-error: false
-            node: 18
-            runs-on: ubuntu-22.04
-            BUILD_TOOLS: ON
-            BUILD_TYPE: Release
-            CCOMPILER: clang-15
-            CXXCOMPILER: clang++-15
-            CUCUMBER_TIMEOUT: 60000
+          # - name: clang-15-release
+          #   continue-on-error: false
+          #   node: 18
+          #   runs-on: ubuntu-22.04
+          #   BUILD_TOOLS: ON
+          #   BUILD_TYPE: Release
+          #   CCOMPILER: clang-15
+          #   CXXCOMPILER: clang++-15
+          #   CUCUMBER_TIMEOUT: 60000
 
-          - name: clang-15-debug
-            continue-on-error: false
-            node: 18
-            runs-on: ubuntu-22.04
-            BUILD_TOOLS: ON
-            BUILD_TYPE: Debug
-            CCOMPILER: clang-15
-            CXXCOMPILER: clang++-15
-            CUCUMBER_TIMEOUT: 60000
+          # - name: clang-15-debug
+          #   continue-on-error: false
+          #   node: 18
+          #   runs-on: ubuntu-22.04
+          #   BUILD_TOOLS: ON
+          #   BUILD_TYPE: Debug
+          #   CCOMPILER: clang-15
+          #   CXXCOMPILER: clang++-15
+          #   CUCUMBER_TIMEOUT: 60000
 
-          - name: clang-18-debug-clang-tidy
-            continue-on-error: false
-            node: 18
-            runs-on: ubuntu-24.04
-            BUILD_TOOLS: ON
-            BUILD_TYPE: Debug
-            CCOMPILER: clang-18
-            CXXCOMPILER: clang++-18
-            CUCUMBER_TIMEOUT: 60000
-            ENABLE_CLANG_TIDY: ON
+          # - name: clang-18-debug-clang-tidy
+          #   continue-on-error: false
+          #   node: 18
+          #   runs-on: ubuntu-24.04
+          #   BUILD_TOOLS: ON
+          #   BUILD_TYPE: Debug
+          #   CCOMPILER: clang-18
+          #   CXXCOMPILER: clang++-18
+          #   CUCUMBER_TIMEOUT: 60000
+          #   ENABLE_CLANG_TIDY: ON
 
 
-          - name: clang-14-release
-            continue-on-error: false
-            node: 18
-            runs-on: ubuntu-22.04
-            BUILD_TOOLS: ON
-            BUILD_TYPE: Release
-            CCOMPILER: clang-14
-            CXXCOMPILER: clang++-14
-            CUCUMBER_TIMEOUT: 60000
+          # - name: clang-14-release
+          #   continue-on-error: false
+          #   node: 18
+          #   runs-on: ubuntu-22.04
+          #   BUILD_TOOLS: ON
+          #   BUILD_TYPE: Release
+          #   CCOMPILER: clang-14
+          #   CXXCOMPILER: clang++-14
+          #   CUCUMBER_TIMEOUT: 60000
 
-          - name: clang-13-release
-            continue-on-error: false
-            node: 18
-            runs-on: ubuntu-22.04
-            BUILD_TOOLS: ON
-            BUILD_TYPE: Release
-            CCOMPILER: clang-13
-            CXXCOMPILER: clang++-13
-            CUCUMBER_TIMEOUT: 60000
+          # - name: clang-13-release
+          #   continue-on-error: false
+          #   node: 18
+          #   runs-on: ubuntu-22.04
+          #   BUILD_TOOLS: ON
+          #   BUILD_TYPE: Release
+          #   CCOMPILER: clang-13
+          #   CXXCOMPILER: clang++-13
+          #   CUCUMBER_TIMEOUT: 60000
 
           - name: conan-linux-debug-asan-ubsan
             continue-on-error: false