--- a/.github/workflows/osrm-backend.yml
+++ b/.github/workflows/osrm-backend.yml
@@ -535,21 +535,21 @@ jobs:
           echo "PKG_CONFIG_PATH=${OSRM_INSTALL_DIR}/lib/pkgconfig" >> $GITHUB_ENV
         fi
         popd
-    - name: Build example
-      if: ${{ matrix.NODE_PACKAGE_TESTS_ONLY != 'ON' }}
-      run: |
-        mkdir example/build && pushd example/build
-        cmake .. -DCMAKE_BUILD_TYPE=${BUILD_TYPE}
-        make --jobs=${JOBS}
-        popd
+    # - name: Build example
+    #   if: ${{ matrix.NODE_PACKAGE_TESTS_ONLY != 'ON' }}
+    #   run: |
+    #     mkdir example/build && pushd example/build
+    #     cmake .. -DCMAKE_BUILD_TYPE=${BUILD_TYPE}
+    #     make --jobs=${JOBS}
+    #     popd
     - name: Run all tests
       if: ${{ matrix.NODE_PACKAGE_TESTS_ONLY != 'ON' }}
       run: |
         make -C test/data benchmark
 
-        # macOS SIP strips the linker path. Reset this inside the running shell
-        export LD_LIBRARY_PATH=${{ env.LD_LIBRARY_PATH }}
-        ./example/build/osrm-example test/data/mld/monaco.osrm
+        # # macOS SIP strips the linker path. Reset this inside the running shell
+        # export LD_LIBRARY_PATH=${{ env.LD_LIBRARY_PATH }}
+        # ./example/build/osrm-example test/data/mld/monaco.osrm
 
         # All tests assume to be run from the build directory
         pushd ${OSRM_BUILD_DIR}