--- a/src/extractor/extractor_callbacks.cpp
+++ b/src/extractor/extractor_callbacks.cpp
@@ -23,8 +23,6 @@
 #include <vector>
 
 #ifdef _MSC_VER
-#if (_MSC_VER >= 1928)
-#ifdef _DEBUG
 namespace osrm
 {
 namespace extractor
@@ -36,8 +34,6 @@ const ByEdgeOrByMeterValue::ValueByMeter ByEdgeOrByMeterValue::by_meter;
 } // namespace detail
 } // namespace extractor
 } // namespace osrm
-#endif
-#endif
 #endif
 
 namespace osrm::extractor