Initially I wanted to enable it by default, but I faced a lot of issues while trying to make it working on all CI jobs, so I decided to enable it only on "release" jobs, i.e. ones which produce NodeJS binaries + on benchmarks job.







<!-- BENCHMARK_RESULTS_START -->
<details><summary><h2>Benchmark Results</h2></summary>

| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1046.01<br/>plain u32: 1082.19<br/>aliased double: 1082.65<br/>plain double: 1140.65 | aliased u32: 1062.51<br/>plain u32: 1056.07<br/>aliased double: 932.948<br/>plain double: 894.062 |
| e2e_match_ch | Ops: 47.14 ± 0.14 ops/s. Best: 47.40 ops/s<br/>Total: 2778.84ms ± 8.56ms. Best: 2763.82ms<br/>Min time: 2.05ms ± 0.05ms<br/>Mean time: 21.21ms ± 0.07ms<br/>Median time: 16.20ms ± 0.14ms<br/>95th percentile: 68.12ms ± 0.76ms<br/>99th percentile: 83.23ms ± 0.41ms<br/>Max time: 92.18ms ± 1.00ms | Ops: 47.21 ± 0.16 ops/s. Best: 47.54 ops/s<br/>Total: 2774.40ms ± 9.48ms. Best: 2755.36ms<br/>Min time: 2.10ms ± 0.06ms<br/>Mean time: 21.18ms ± 0.07ms<br/>Median time: 16.83ms ± 0.15ms<br/>95th percentile: 68.03ms ± 0.61ms<br/>99th percentile: 81.90ms ± 0.97ms<br/>Max time: 88.16ms ± 1.65ms |
| e2e_match_mld | Ops: 62.88 ± 0.22 ops/s. Best: 63.34 ops/s<br/>Total: 2083.78ms ± 7.47ms. Best: 2068.19ms<br/>Min time: 1.74ms ± 0.06ms<br/>Mean time: 15.90ms ± 0.06ms<br/>Median time: 8.26ms ± 0.11ms<br/>95th percentile: 52.69ms ± 0.50ms<br/>99th percentile: 60.03ms ± 0.64ms<br/>Max time: 70.17ms ± 1.00ms | Ops: 66.77 ± 0.23 ops/s. Best: 67.01 ops/s<br/>Total: 1961.62ms ± 6.80ms. Best: 1954.98ms<br/>Min time: 1.74ms ± 0.03ms<br/>Mean time: 14.98ms ± 0.05ms<br/>Median time: 8.04ms ± 0.08ms<br/>95th percentile: 49.47ms ± 0.44ms<br/>99th percentile: 56.92ms ± 0.82ms<br/>Max time: 67.10ms ± 0.96ms |
| e2e_nearest_ch | Ops: 820.87 ± 7.22 ops/s. Best: 828.65 ops/s<br/>Total: 1218.44ms ± 10.82ms. Best: 1206.79ms<br/>Min time: 0.98ms ± 0.01ms<br/>Mean time: 1.22ms ± 0.01ms<br/>Median time: 1.14ms ± 0.01ms<br/>95th percentile: 1.63ms ± 0.01ms<br/>99th percentile: 1.68ms ± 0.02ms<br/>Max time: 3.86ms ± 2.10ms | Ops: 826.89 ± 9.30 ops/s. Best: 842.79 ops/s<br/>Total: 1209.43ms ± 14.51ms. Best: 1186.54ms<br/>Min time: 0.98ms ± 0.01ms<br/>Mean time: 1.21ms ± 0.02ms<br/>Median time: 1.13ms ± 0.01ms<br/>95th percentile: 1.61ms ± 0.01ms<br/>99th percentile: 1.67ms ± 0.02ms<br/>Max time: 3.90ms ± 2.11ms |
| e2e_nearest_mld | Ops: 817.03 ± 6.41 ops/s. Best: 827.38 ops/s<br/>Total: 1223.57ms ± 9.25ms. Best: 1208.63ms<br/>Min time: 0.99ms ± 0.01ms<br/>Mean time: 1.22ms ± 0.01ms<br/>Median time: 1.14ms ± 0.01ms<br/>95th percentile: 1.63ms ± 0.01ms<br/>99th percentile: 1.69ms ± 0.02ms<br/>Max time: 4.14ms ± 2.43ms | Ops: 818.24 ± 3.30 ops/s. Best: 823.95 ops/s<br/>Total: 1222.30ms ± 4.78ms. Best: 1213.66ms<br/>Min time: 0.99ms ± 0.01ms<br/>Mean time: 1.22ms ± 0.00ms<br/>Median time: 1.14ms ± 0.01ms<br/>95th percentile: 1.62ms ± 0.01ms<br/>99th percentile: 1.68ms ± 0.00ms<br/>Max time: 4.31ms ± 2.51ms |
| e2e_route_ch | Ops: 360.96 ± 3.42 ops/s. Best: 367.87 ops/s<br/>Total: 2770.47ms ± 26.48ms. Best: 2718.32ms<br/>Min time: 1.21ms ± 0.02ms<br/>Mean time: 2.77ms ± 0.03ms<br/>Median time: 2.79ms ± 0.03ms<br/>95th percentile: 3.66ms ± 0.03ms<br/>99th percentile: 4.04ms ± 0.05ms<br/>Max time: 6.25ms ± 1.84ms | Ops: 367.12 ± 1.79 ops/s. Best: 370.85 ops/s<br/>Total: 2724.57ms ± 13.26ms. Best: 2696.53ms<br/>Min time: 1.21ms ± 0.04ms<br/>Mean time: 2.72ms ± 0.01ms<br/>Median time: 2.74ms ± 0.02ms<br/>95th percentile: 3.60ms ± 0.04ms<br/>99th percentile: 3.97ms ± 0.05ms<br/>Max time: 6.23ms ± 1.75ms |
| e2e_route_mld | Ops: 304.30 ± 1.42 ops/s. Best: 306.01 ops/s<br/>Total: 3286.29ms ± 16.18ms. Best: 3267.85ms<br/>Min time: 1.23ms ± 0.03ms<br/>Mean time: 3.29ms ± 0.02ms<br/>Median time: 3.32ms ± 0.02ms<br/>95th percentile: 4.46ms ± 0.01ms<br/>99th percentile: 4.91ms ± 0.05ms<br/>Max time: 7.37ms ± 1.92ms | Ops: 307.85 ± 1.12 ops/s. Best: 309.36 ops/s<br/>Total: 3248.62ms ± 11.85ms. Best: 3232.49ms<br/>Min time: 1.22ms ± 0.02ms<br/>Mean time: 3.25ms ± 0.01ms<br/>Median time: 3.28ms ± 0.02ms<br/>95th percentile: 4.41ms ± 0.02ms<br/>99th percentile: 4.82ms ± 0.04ms<br/>Max time: 7.01ms ± 1.59ms |
| e2e_table_ch | Ops: 315.94 ± 1.60 ops/s. Best: 318.52 ops/s<br/>Total: 3165.47ms ± 15.98ms. Best: 3139.54ms<br/>Min time: 1.60ms ± 0.04ms<br/>Mean time: 3.17ms ± 0.02ms<br/>Median time: 3.16ms ± 0.02ms<br/>95th percentile: 4.38ms ± 0.03ms<br/>99th percentile: 4.72ms ± 0.06ms<br/>Max time: 7.38ms ± 2.61ms | Ops: 317.92 ± 1.19 ops/s. Best: 319.83 ops/s<br/>Total: 3145.81ms ± 11.64ms. Best: 3126.67ms<br/>Min time: 1.59ms ± 0.02ms<br/>Mean time: 3.15ms ± 0.01ms<br/>Median time: 3.15ms ± 0.01ms<br/>95th percentile: 4.35ms ± 0.03ms<br/>99th percentile: 4.70ms ± 0.04ms<br/>Max time: 6.98ms ± 2.24ms |
| e2e_table_mld | Ops: 115.13 ± 0.29 ops/s. Best: 115.64 ops/s<br/>Total: 8685.42ms ± 22.88ms. Best: 8647.75ms<br/>Min time: 3.42ms ± 0.08ms<br/>Mean time: 8.69ms ± 0.02ms<br/>Median time: 8.65ms ± 0.04ms<br/>95th percentile: 13.34ms ± 0.07ms<br/>99th percentile: 14.14ms ± 0.10ms<br/>Max time: 16.26ms ± 1.41ms | Ops: 116.44 ± 0.20 ops/s. Best: 116.82 ops/s<br/>Total: 8589.06ms ± 14.89ms. Best: 8560.37ms<br/>Min time: 3.46ms ± 0.11ms<br/>Mean time: 8.59ms ± 0.01ms<br/>Median time: 8.54ms ± 0.04ms<br/>95th percentile: 13.13ms ± 0.06ms<br/>99th percentile: 14.02ms ± 0.10ms<br/>Max time: 16.17ms ± 1.62ms |
| e2e_trip_ch | Ops: 101.42 ± 0.14 ops/s. Best: 101.57 ops/s<br/>Total: 9859.95ms ± 12.83ms. Best: 9844.99ms<br/>Min time: 1.52ms ± 0.16ms<br/>Mean time: 9.86ms ± 0.01ms<br/>Median time: 9.33ms ± 0.03ms<br/>95th percentile: 17.57ms ± 0.08ms<br/>99th percentile: 19.33ms ± 0.11ms<br/>Max time: 20.81ms ± 0.07ms | Ops: 103.88 ± 0.24 ops/s. Best: 104.31 ops/s<br/>Total: 9625.91ms ± 22.50ms. Best: 9587.11ms<br/>Min time: 1.46ms ± 0.11ms<br/>Mean time: 9.63ms ± 0.02ms<br/>Median time: 9.12ms ± 0.02ms<br/>95th percentile: 17.25ms ± 0.05ms<br/>99th percentile: 18.60ms ± 0.09ms<br/>Max time: 19.82ms ± 0.17ms |
| e2e_trip_mld | Ops: 61.29 ± 0.16 ops/s. Best: 61.57 ops/s<br/>Total: 16313.46ms ± 42.77ms. Best: 16240.65ms<br/>Min time: 1.57ms ± 0.19ms<br/>Mean time: 16.32ms ± 0.04ms<br/>Median time: 15.88ms ± 0.03ms<br/>95th percentile: 26.90ms ± 0.08ms<br/>99th percentile: 28.67ms ± 0.17ms<br/>Max time: 32.02ms ± 1.85ms | Ops: 62.60 ± 0.29 ops/s. Best: 62.97 ops/s<br/>Total: 15974.19ms ± 75.13ms. Best: 15880.60ms<br/>Min time: 1.45ms ± 0.17ms<br/>Mean time: 15.97ms ± 0.08ms<br/>Median time: 15.55ms ± 0.12ms<br/>95th percentile: 26.33ms ± 0.15ms<br/>99th percentile: 27.81ms ± 0.21ms<br/>Max time: 31.13ms ± 2.15ms |
| json-render | String: 6.40608ms<br/>Stringstream: 9.76474ms<br/>Vector: 6.06719ms | String: 5.80294ms<br/>Stringstream: 9.39139ms<br/>Vector: 6.26933ms |
| match_ch | Default radius: <br/>4.43682ms/req at 82 coordinate<br/>0.0541076ms/coordinate<br/>Radius 10m: <br/>15.8013ms/req at 82 coordinate<br/>0.192699ms/coordinate | Default radius: <br/>4.45801ms/req at 82 coordinate<br/>0.054366ms/coordinate<br/>Radius 10m: <br/>15.2671ms/req at 82 coordinate<br/>0.186184ms/coordinate |
| match_mld | Default radius: <br/>3.30351ms/req at 82 coordinate<br/>0.0402867ms/coordinate<br/>Radius 10m: <br/>11.2358ms/req at 82 coordinate<br/>0.137022ms/coordinate | Default radius: <br/>2.96238ms/req at 82 coordinate<br/>0.0361265ms/coordinate<br/>Radius 10m: <br/>10.9574ms/req at 82 coordinate<br/>0.133627ms/coordinate |
| osrm_contract | Time: 96.62s Peak RAM: 201.29MB | Time: 94.66s Peak RAM: 203.28MB |
| osrm_customize | Time: 1.31s Peak RAM: 117.07MB | Time: 1.29s Peak RAM: 116.73MB |
| osrm_extract | Time: 11.68s Peak RAM: 413.35MB | Time: 11.69s Peak RAM: 408.01MB |
| osrm_partition | Time: 1.95s Peak RAM: 136.22MB | Time: 1.91s Peak RAM: 133.41MB |
| packedvector | random write:<br/>std::vector 9571.13 ms<br/>util::packed_vector 79499.5 ms<br/>slowdown: 8.30618<br/>random read:<br/>std::vector 8173.47 ms<br/>util::packed_vector 32434.6 ms<br/>slowdown: 3.96828 | random write:<br/>std::vector 10507.4 ms<br/>util::packed_vector 71922.8 ms<br/>slowdown: 6.84494<br/>random read:<br/>std::vector 10770.4 ms<br/>util::packed_vector 29695.8 ms<br/>slowdown: 2.75717 |
| random_match_ch | 500 matches, default radius<br/>ops: 226.34 ± 1.30 ops/s. best: 228.35ops/s.<br/>total: 251.84 ± 1.45ms. best: 249.62ms.<br/>avg: 4.42 ± 0.03ms<br/>min: 0.15 ± 0.01ms<br/>max: 24.66 ± 0.27ms<br/>p99: 24.66 ± 0.27ms<br/><br/>500 matches, radius=10<br/>ops: 68.15 ± 0.47 ops/s. best: 69.28ops/s.<br/>total: 939.17 ± 6.47ms. best: 923.76ms.<br/>avg: 14.67 ± 0.10ms<br/>min: 0.15 ± 0.01ms<br/>max: 234.98 ± 1.20ms<br/>p99: 234.98 ± 1.20ms<br/><br/>500 matches, radius=20<br/>ops: 16.17 ± 0.08 ops/s. best: 16.30ops/s.<br/>total: 4019.82 ± 19.25ms. best: 3988.19ms.<br/>avg: 61.84 ± 0.30ms<br/>min: 0.28 ± 0.01ms<br/>max: 1178.49 ± 5.37ms<br/>p99: 1178.49 ± 5.37ms | 500 matches, default radius<br/>ops: 224.90 ± 1.86 ops/s. best: 227.81ops/s.<br/>total: 253.46 ± 2.07ms. best: 250.21ms.<br/>avg: 4.45 ± 0.04ms<br/>min: 0.15 ± 0.01ms<br/>max: 23.12 ± 0.33ms<br/>p99: 23.12 ± 0.33ms<br/><br/>500 matches, radius=10<br/>ops: 66.61 ± 0.41 ops/s. best: 67.42ops/s.<br/>total: 960.81 ± 5.84ms. best: 949.29ms.<br/>avg: 15.01 ± 0.09ms<br/>min: 0.16 ± 0.01ms<br/>max: 226.33 ± 1.48ms<br/>p99: 226.33 ± 1.48ms<br/><br/>500 matches, radius=20<br/>ops: 15.99 ± 0.05 ops/s. best: 16.09ops/s.<br/>total: 4064.89 ± 12.91ms. best: 4040.73ms.<br/>avg: 62.54 ± 0.20ms<br/>min: 0.31 ± 0.01ms<br/>max: 1158.33 ± 4.72ms<br/>p99: 1158.33 ± 4.72ms |
| random_match_mld | 500 matches, default radius<br/>ops: 309.82 ± 1.95 ops/s. best: 312.38ops/s.<br/>total: 183.99 ± 1.16ms. best: 182.47ms.<br/>avg: 3.23 ± 0.02ms<br/>min: 0.14 ± 0.00ms<br/>max: 19.06 ± 0.36ms<br/>p99: 19.06 ± 0.36ms<br/><br/>500 matches, radius=10<br/>ops: 108.02 ± 0.46 ops/s. best: 108.68ops/s.<br/>total: 592.49 ± 2.57ms. best: 588.87ms.<br/>avg: 9.26 ± 0.04ms<br/>min: 0.15 ± 0.00ms<br/>max: 112.71 ± 1.70ms<br/>p99: 112.71 ± 1.70ms<br/><br/>500 matches, radius=20<br/>ops: 21.59 ± 0.08 ops/s. best: 21.70ops/s.<br/>total: 3011.35 ± 11.74ms. best: 2995.48ms.<br/>avg: 46.33 ± 0.18ms<br/>min: 0.20 ± 0.01ms<br/>max: 595.87 ± 5.77ms<br/>p99: 595.87 ± 5.77ms | 500 matches, default radius<br/>ops: 321.06 ± 4.74 ops/s. best: 330.48ops/s.<br/>total: 177.58 ± 2.60ms. best: 172.48ms.<br/>avg: 3.12 ± 0.05ms<br/>min: 0.14 ± 0.00ms<br/>max: 18.41 ± 0.34ms<br/>p99: 18.41 ± 0.34ms<br/><br/>500 matches, radius=10<br/>ops: 115.17 ± 0.97 ops/s. best: 116.44ops/s.<br/>total: 555.76 ± 4.70ms. best: 549.62ms.<br/>avg: 8.68 ± 0.07ms<br/>min: 0.15 ± 0.01ms<br/>max: 106.49 ± 0.61ms<br/>p99: 106.49 ± 0.61ms<br/><br/>500 matches, radius=20<br/>ops: 23.08 ± 0.09 ops/s. best: 23.24ops/s.<br/>total: 2815.81 ± 11.29ms. best: 2796.56ms.<br/>avg: 43.32 ± 0.17ms<br/>min: 0.19 ± 0.01ms<br/>max: 556.04 ± 2.87ms<br/>p99: 556.04 ± 2.87ms |
| random_nearest_ch | 10000 nearest, number_of_results=1<br/>ops: 24506.51 ± 186.30 ops/s. best: 24713.49ops/s.<br/>total: 408.08 ± 3.09ms. best: 404.64ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.15 ± 0.02ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 18932.33 ± 169.82 ops/s. best: 19136.38ops/s.<br/>total: 528.26 ± 4.73ms. best: 522.56ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14995.26 ± 50.54 ops/s. best: 15063.96ops/s.<br/>total: 666.89 ± 2.25ms. best: 663.84ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.00ms<br/>p99: 0.13 ± 0.00ms | 10000 nearest, number_of_results=1<br/>ops: 24366.64 ± 158.51 ops/s. best: 24610.10ops/s.<br/>total: 410.42 ± 2.68ms. best: 406.34ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.15 ± 0.02ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 18956.11 ± 153.19 ops/s. best: 19201.78ops/s.<br/>total: 527.58 ± 4.25ms. best: 520.79ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.01ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 15267.55 ± 100.47 ops/s. best: 15462.02ops/s.<br/>total: 655.03 ± 4.30ms. best: 646.75ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.00ms<br/>p99: 0.13 ± 0.00ms |
| random_nearest_mld | 10000 nearest, number_of_results=1<br/>ops: 24436.26 ± 268.31 ops/s. best: 24922.99ops/s.<br/>total: 409.29 ± 4.48ms. best: 401.24ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.15 ± 0.02ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 19089.23 ± 67.90 ops/s. best: 19200.42ops/s.<br/>total: 523.86 ± 1.89ms. best: 520.82ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.01ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 15126.00 ± 100.06 ops/s. best: 15265.11ops/s.<br/>total: 661.15 ± 4.39ms. best: 655.09ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.00ms<br/>p99: 0.13 ± 0.00ms | 10000 nearest, number_of_results=1<br/>ops: 24422.63 ± 271.63 ops/s. best: 24776.17ops/s.<br/>total: 409.52 ± 4.57ms. best: 403.61ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.15 ± 0.02ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 18712.77 ± 95.51 ops/s. best: 18818.73ops/s.<br/>total: 534.41 ± 2.71ms. best: 531.39ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 15009.78 ± 169.28 ops/s. best: 15338.43ops/s.<br/>total: 666.33 ± 6.94ms. best: 651.96ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.00ms<br/>p99: 0.13 ± 0.00ms |
| random_route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 546.90 ± 1.36 ops/s. best: 549.11ops/s.<br/>total: 1799.26 ± 4.46ms. best: 1792.00ms.<br/>avg: 1.83 ± 0.00ms<br/>min: 0.31 ± 0.01ms<br/>max: 3.08 ± 0.18ms<br/>p99: 2.66 ± 0.01ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 642.95 ± 2.89 ops/s. best: 647.50ops/s.<br/>total: 1555.37 ± 6.97ms. best: 1544.41ms.<br/>avg: 1.56 ± 0.01ms<br/>min: 0.05 ± 0.00ms<br/>max: 4.49 ± 0.08ms<br/>p99: 3.35 ± 0.02ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 1114.95 ± 6.97 ops/s. best: 1129.87ops/s.<br/>total: 882.60 ± 5.48ms. best: 870.90ms.<br/>avg: 0.90 ± 0.01ms<br/>min: 0.22 ± 0.01ms<br/>max: 1.39 ± 0.03ms<br/>p99: 1.26 ± 0.01ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 1234.30 ± 2.73 ops/s. best: 1237.89ops/s.<br/>total: 810.18 ± 1.79ms. best: 807.83ms.<br/>avg: 0.81 ± 0.00ms<br/>min: 0.04 ± 0.00ms<br/>max: 3.66 ± 0.07ms<br/>p99: 2.00 ± 0.01ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 553.86 ± 6.04 ops/s. best: 559.60ops/s.<br/>total: 1776.93 ± 19.96ms. best: 1758.39ms.<br/>avg: 1.81 ± 0.02ms<br/>min: 0.31 ± 0.01ms<br/>max: 3.05 ± 0.23ms<br/>p99: 2.62 ± 0.07ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 629.17 ± 1.68 ops/s. best: 631.46ops/s.<br/>total: 1589.41 ± 4.23ms. best: 1583.63ms.<br/>avg: 1.59 ± 0.00ms<br/>min: 0.05 ± 0.00ms<br/>max: 3.75 ± 0.07ms<br/>p99: 3.49 ± 0.02ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 1090.58 ± 7.65 ops/s. best: 1105.79ops/s.<br/>total: 902.33 ± 6.28ms. best: 889.86ms.<br/>avg: 0.92 ± 0.01ms<br/>min: 0.22 ± 0.01ms<br/>max: 1.45 ± 0.01ms<br/>p99: 1.29 ± 0.02ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 1154.01 ± 7.73 ops/s. best: 1163.73ops/s.<br/>total: 866.60 ± 5.83ms. best: 859.31ms.<br/>avg: 0.87 ± 0.01ms<br/>min: 0.04 ± 0.00ms<br/>max: 2.26 ± 0.13ms<br/>p99: 2.08 ± 0.01ms |
| random_route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 268.52 ± 1.44 ops/s. best: 271.62ops/s.<br/>total: 3664.63 ± 19.52ms. best: 3622.76ms.<br/>avg: 3.72 ± 0.02ms<br/>min: 0.32 ± 0.01ms<br/>max: 8.24 ± 0.14ms<br/>p99: 6.25 ± 0.09ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 263.66 ± 0.78 ops/s. best: 265.56ops/s.<br/>total: 3792.76 ± 11.09ms. best: 3765.58ms.<br/>avg: 3.79 ± 0.01ms<br/>min: 0.05 ± 0.00ms<br/>max: 8.56 ± 0.12ms<br/>p99: 7.69 ± 0.06ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 356.70 ± 2.63 ops/s. best: 359.75ops/s.<br/>total: 2758.84 ± 20.45ms. best: 2735.21ms.<br/>avg: 2.80 ± 0.02ms<br/>min: 0.28 ± 0.00ms<br/>max: 6.99 ± 0.09ms<br/>p99: 4.90 ± 0.05ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 326.53 ± 0.76 ops/s. best: 327.50ops/s.<br/>total: 3062.50 ± 7.02ms. best: 3053.48ms.<br/>avg: 3.06 ± 0.01ms<br/>min: 0.04 ± 0.00ms<br/>max: 6.90 ± 0.24ms<br/>p99: 6.09 ± 0.04ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 272.13 ± 2.02 ops/s. best: 275.16ops/s.<br/>total: 3616.16 ± 27.03ms. best: 3576.14ms.<br/>avg: 3.67 ± 0.03ms<br/>min: 0.32 ± 0.01ms<br/>max: 8.12 ± 0.18ms<br/>p99: 6.12 ± 0.07ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 266.38 ± 1.41 ops/s. best: 269.33ops/s.<br/>total: 3754.22 ± 19.78ms. best: 3712.96ms.<br/>avg: 3.75 ± 0.02ms<br/>min: 0.05 ± 0.00ms<br/>max: 8.50 ± 0.08ms<br/>p99: 7.50 ± 0.09ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 357.49 ± 1.74 ops/s. best: 361.34ops/s.<br/>total: 2752.58 ± 13.31ms. best: 2723.19ms.<br/>avg: 2.80 ± 0.01ms<br/>min: 0.28 ± 0.00ms<br/>max: 6.79 ± 0.15ms<br/>p99: 4.80 ± 0.06ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 328.97 ± 2.90 ops/s. best: 333.07ops/s.<br/>total: 3040.06 ± 26.72ms. best: 3002.38ms.<br/>avg: 3.04 ± 0.03ms<br/>min: 0.04 ± 0.00ms<br/>max: 6.66 ± 0.09ms<br/>p99: 5.98 ± 0.07ms |
| random_table_ch | 250 tables, 3 coordinates<br/>ops: 1511.47 ± 28.35 ops/s. best: 1558.59ops/s.<br/>total: 165.46 ± 3.07ms. best: 160.40ms.<br/>avg: 0.66 ± 0.01ms<br/>min: 0.43 ± 0.01ms<br/>max: 1.00 ± 0.28ms<br/>p99: 0.82 ± 0.03ms<br/><br/>250 tables, 25 coordinates<br/>ops: 179.77 ± 1.15 ops/s. best: 181.91ops/s.<br/>total: 1390.74 ± 8.84ms. best: 1374.32ms.<br/>avg: 5.56 ± 0.04ms<br/>min: 4.82 ± 0.06ms<br/>max: 6.22 ± 0.06ms<br/>p99: 6.14 ± 0.04ms<br/><br/>250 tables, 50 coordinates<br/>ops: 88.14 ± 0.69 ops/s. best: 89.11ops/s.<br/>total: 2836.59 ± 22.28ms. best: 2805.54ms.<br/>avg: 11.35 ± 0.09ms<br/>min: 10.20 ± 0.18ms<br/>max: 12.50 ± 0.14ms<br/>p99: 12.29 ± 0.06ms | 250 tables, 3 coordinates<br/>ops: 1480.31 ± 12.95 ops/s. best: 1501.85ops/s.<br/>total: 168.90 ± 1.47ms. best: 166.46ms.<br/>avg: 0.68 ± 0.01ms<br/>min: 0.48 ± 0.02ms<br/>max: 1.01 ± 0.21ms<br/>p99: 0.84 ± 0.02ms<br/><br/>250 tables, 25 coordinates<br/>ops: 176.25 ± 1.67 ops/s. best: 178.16ops/s.<br/>total: 1418.63 ± 13.36ms. best: 1403.20ms.<br/>avg: 5.67 ± 0.05ms<br/>min: 4.97 ± 0.10ms<br/>max: 6.67 ± 0.30ms<br/>p99: 6.29 ± 0.02ms<br/><br/>250 tables, 50 coordinates<br/>ops: 86.19 ± 0.62 ops/s. best: 86.96ops/s.<br/>total: 2900.64 ± 20.85ms. best: 2874.94ms.<br/>avg: 11.60 ± 0.08ms<br/>min: 10.47 ± 0.11ms<br/>max: 12.71 ± 0.08ms<br/>p99: 12.52 ± 0.12ms |
…