@SiarheiFedartsou 2024-06-29
> Looks good. How about making it the default? The few platforms that don't support it may disable it explicitly. Or is LTO support still not widespread enough?

Well, "few platforms" in our case is ~50% of our CI jobs if I remember correctly 😄 But in general we indeed can make it default and just disable it on jobs where it doesn't work yet - in general it should be possible to make it working anywhere...